Rename and move visitor into subpackage
